Zarr: split code into multiple files
