+++ gdal/frmts/zarr/zarr.h
@@ -1,0 +1,359 @@
+/******************************************************************************
+ *
+ * Project:  GDAL
+ * Purpose:  Zarr driver
+ * Author:   Even Rouault <even dot rouault at spatialys.com>
+ *
+ ******************************************************************************
+ * Copyright (c) 2021, Even Rouault <even dot rouault at spatialys.com>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+ * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
+ ****************************************************************************/
+
+#ifndef ZARR_H
+#define ZARR_H
+
+#include "cpl_compressor.h"
+#include "cpl_json.h"
+#include "gdal_priv.h"
+#include "memmultidim.h"
+
+/************************************************************************/
+/*                            ZarrDataset                               */
+/************************************************************************/
+
+class ZarrDataset final: public GDALDataset
+{
+    std::shared_ptr<GDALGroup> m_poRootGroup{};
+    CPLStringList              m_aosSubdatasets{};
+
+    static GDALDataset* OpenMultidim(const char* pszFilename,
+                                     bool bUpdateMode,
+                                     CSLConstList papszOpenOptions);
+
+public:
+    static int Identify( GDALOpenInfo *poOpenInfo );
+    static GDALDataset* Open(GDALOpenInfo* poOpenInfo);
+    static GDALDataset* CreateMultiDimensional( const char * pszFilename,
+                                                CSLConstList /*papszRootGroupOptions*/,
+                                                CSLConstList /*papszOptions*/ );
+
+    const char* GetMetadataItem(const char* pszName, const char* pszDomain) override;
+    char** GetMetadata(const char* pszDomain) override;
+
+    std::shared_ptr<GDALGroup> GetRootGroup() const override { return m_poRootGroup; }
+};
+
+/************************************************************************/
+/*                        ZarrAttributeGroup()                          */
+/************************************************************************/
+
+class ZarrAttributeGroup
+{
+    // Use a MEMGroup as a convenient container for attributes.
+    MEMGroup m_oGroup;
+    bool m_bModified = false;
+
+public:
+    explicit ZarrAttributeGroup(const std::string& osParentName);
+
+    void Init(const CPLJSONObject& obj, bool bUpdatable);
+
+    std::shared_ptr<GDALAttribute> GetAttribute(const std::string& osName) const { return m_oGroup.GetAttribute(osName); }
+
+    std::vector<std::shared_ptr<GDALAttribute>> GetAttributes(CSLConstList papszOptions = nullptr) const { return m_oGroup.GetAttributes(papszOptions); }
+
+    std::shared_ptr<GDALAttribute> CreateAttribute(
+        const std::string& osName,
+        const std::vector<GUInt64>& anDimensions,
+        const GDALExtendedDataType& oDataType,
+        CSLConstList /* papszOptions */ = nullptr )
+    {
+        auto poAttr = m_oGroup.CreateAttribute(osName, anDimensions, oDataType, nullptr);
+        if( poAttr )
+        {
+            m_bModified = true;
+        }
+        return poAttr;
+    }
+
+    void SetUpdatable(bool bUpdatable)
+    {
+        auto attrs = m_oGroup.GetAttributes(nullptr);
+        for( auto attr: attrs )
+        {
+            auto memAttr = std::dynamic_pointer_cast<MEMAttribute>(attr);
+            if( memAttr )
+                memAttr->SetWritable(bUpdatable);
+        }
+    }
+
+    bool IsModified() const
+    {
+        if( m_bModified )
+            return true;
+        const auto attrs = m_oGroup.GetAttributes(nullptr);
+        for( const auto& attr: attrs )
+        {
+            const auto memAttr = std::dynamic_pointer_cast<MEMAttribute>(attr);
+            if( memAttr && memAttr->IsModified() )
+                return true;
+        }
+        return false;
+    }
+
+    CPLJSONObject Serialize() const;
+};
+
+/************************************************************************/
+/*                             ZarrGroup                                */
+/************************************************************************/
+
+class ZarrArray;
+
+class ZarrGroupBase CPL_NON_FINAL: public GDALGroup
+{
+protected:
+    std::string m_osDirectoryName{};
+    mutable std::map<CPLString, std::shared_ptr<GDALGroup>> m_oMapGroups{};
+    mutable std::map<CPLString, std::shared_ptr<GDALMDArray>> m_oMapMDArrays{};
+    mutable std::map<CPLString, std::shared_ptr<GDALDimensionWeakIndexingVar>> m_oMapDimensions{};
+    mutable bool m_bDirectoryExplored = false;
+    mutable std::vector<std::string> m_aosGroups{};
+    mutable std::vector<std::string> m_aosArrays{};
+    mutable ZarrAttributeGroup                        m_oAttrGroup;
+    mutable bool                                      m_bAttributesLoaded = false;
+    bool                                              m_bReadFromZMetadata = false;
+    mutable bool                                      m_bDimensionsInstantiated = false;
+    bool                                              m_bUpdatable = false;
+
+    virtual void ExploreDirectory() const = 0;
+    virtual void LoadAttributes() const = 0;
+
+public:
+    ZarrGroupBase(const std::string& osParentName, const std::string& osName):
+        GDALGroup(osParentName, osName),
+        m_oAttrGroup(osParentName) {}
+
+    std::shared_ptr<GDALAttribute> GetAttribute(const std::string& osName) const override
+        { LoadAttributes(); return m_oAttrGroup.GetAttribute(osName); }
+
+    std::vector<std::shared_ptr<GDALAttribute>> GetAttributes(CSLConstList papszOptions = nullptr) const override
+        { LoadAttributes(); return m_oAttrGroup.GetAttributes(papszOptions); }
+
+    std::shared_ptr<GDALAttribute> CreateAttribute(
+        const std::string& osName,
+        const std::vector<GUInt64>& anDimensions,
+        const GDALExtendedDataType& oDataType,
+        CSLConstList papszOptions = nullptr) override;
+
+    std::vector<std::shared_ptr<GDALDimension>> GetDimensions(CSLConstList papszOptions) const override;
+
+    std::vector<std::string> GetMDArrayNames(CSLConstList papszOptions = nullptr) const override;
+
+    std::vector<std::string> GetGroupNames(CSLConstList papszOptions = nullptr) const override;
+
+    void SetDirectoryName(const std::string& osDirectoryName) { m_osDirectoryName = osDirectoryName; }
+
+    std::shared_ptr<ZarrArray> LoadArray(const std::string& osArrayName,
+                                         const std::string& osZarrayFilename,
+                                         const CPLJSONObject& oRoot,
+                                         bool bLoadedFromZMetadata,
+                                         const CPLJSONObject& oAttributes) const;
+    void RegisterArray(const std::shared_ptr<GDALMDArray>& array) const;
+
+    void SetUpdatable(bool bUpdatable) { m_bUpdatable = bUpdatable; }
+};
+
+class ZarrGroupV2 final: public ZarrGroupBase
+{
+    void ExploreDirectory() const override;
+    void LoadAttributes() const override;
+
+    std::shared_ptr<ZarrGroupV2> GetOrCreateSubGroup(
+                                        const std::string& osSubGroupFullname);
+
+public:
+    ZarrGroupV2(const std::string& osParentName, const std::string& osName):
+        ZarrGroupBase(osParentName, osName) {}
+
+    ~ZarrGroupV2() override;
+
+    static std::shared_ptr<ZarrGroupV2> CreateOnDisk(const std::string& osParentName,
+                                                     const std::string& osName,
+                                                     const std::string& osDirectoryName);
+
+    std::shared_ptr<GDALMDArray> OpenMDArray(const std::string& osName,
+                                             CSLConstList papszOptions = nullptr) const override;
+
+    std::shared_ptr<GDALGroup> OpenGroup(const std::string& osName,
+                                         CSLConstList papszOptions) const override;
+
+    std::shared_ptr<GDALGroup> CreateGroup(const std::string& osName,
+                                           CSLConstList papszOptions = nullptr) override;
+
+    void InitFromZMetadata(const CPLJSONObject& oRoot);
+};
+
+class ZarrGroupV3 final: public ZarrGroupBase
+{
+    void ExploreDirectory() const override;
+    void LoadAttributes() const override;
+
+public:
+    ZarrGroupV3(const std::string& osParentName, const std::string& osName):
+        ZarrGroupBase(osParentName, osName) {}
+
+    std::shared_ptr<GDALMDArray> OpenMDArray(const std::string& osName,
+                                             CSLConstList papszOptions = nullptr) const override;
+
+    std::shared_ptr<GDALGroup> OpenGroup(const std::string& osName,
+                                         CSLConstList papszOptions) const override;
+};
+
+/************************************************************************/
+/*                              DtypeElt()                              */
+/************************************************************************/
+
+struct DtypeElt
+{
+    enum class NativeType
+    {
+        BOOLEAN,
+        UNSIGNED_INT,
+        SIGNED_INT,
+        IEEEFP,
+        COMPLEX_IEEEFP,
+        STRING,
+    };
+
+    NativeType           nativeType = NativeType::BOOLEAN;
+    size_t               nativeOffset = 0;
+    size_t               nativeSize = 0;
+    bool                 needByteSwapping = false;
+    bool                 gdalTypeIsApproxOfNative = false;
+    GDALExtendedDataType gdalType = GDALExtendedDataType::Create(GDT_Unknown);
+    size_t               gdalOffset = 0;
+    size_t               gdalSize = 0;
+};
+
+/************************************************************************/
+/*                             ZarrArray                                */
+/************************************************************************/
+
+class ZarrArray final: public GDALMDArray
+{
+    const std::vector<std::shared_ptr<GDALDimension>> m_aoDims;
+    const GDALExtendedDataType                        m_oType;
+    const std::vector<DtypeElt>                       m_aoDtypeElts;
+    const std::vector<GUInt64>                        m_anBlockSize;
+    GByte                                            *m_pabyNoData = nullptr;
+    std::string                                       m_osDimSeparator { "." };
+    std::string                                       m_osFilename{};
+    mutable std::vector<GByte>                        m_abyRawTileData{};
+    mutable std::vector<GByte>                        m_abyDecodedTileData{};
+    mutable std::vector<uint64_t>                     m_anCachedTiledIndices{};
+    mutable bool                                      m_bCachedTiledValid = false;
+    mutable bool                                      m_bCachedTiledEmpty = false;
+    bool                                              m_bUseOptimizedCodePaths = true;
+    bool                                              m_bFortranOrder = false;
+    const CPLCompressor                              *m_psDecompressor = nullptr;
+    mutable std::vector<GByte>                        m_abyTmpRawTileData{}; // used for Fortran order
+    mutable ZarrAttributeGroup                        m_oAttrGroup;
+    mutable std::shared_ptr<OGRSpatialReference>      m_poSRS{};
+    mutable bool                                      m_bAllocateWorkingBuffersDone = false;
+    mutable bool                                      m_bWorkingBuffersOK = false;
+    std::string                                       m_osRootDirectoryName{};
+    int                                               m_nVersion = 0;
+    bool                                              m_bUpdatable = false;
+
+    ZarrArray(const std::string& osParentName,
+              const std::string& osName,
+              const std::vector<std::shared_ptr<GDALDimension>>& aoDims,
+              const GDALExtendedDataType& oType,
+              const std::vector<DtypeElt>& aoDtypeElts,
+              const std::vector<GUInt64>& anBlockSize,
+              bool bFortranOrder);
+
+    bool LoadTileData(const std::vector<uint64_t>& tileIndices,
+                      bool& bMissingTileOut) const;
+    void BlockTranspose(const std::vector<GByte>& abySrc,
+                        std::vector<GByte>& abyDst) const;
+
+    bool AllocateWorkingBuffers() const;
+
+protected:
+    bool IRead(const GUInt64* arrayStartIdx,
+                      const size_t* count,
+                      const GInt64* arrayStep,
+                      const GPtrDiff_t* bufferStride,
+                      const GDALExtendedDataType& bufferDataType,
+                      void* pDstBuffer) const override;
+
+    bool IsCacheable() const override { return false; }
+
+public:
+    ~ZarrArray() override;
+
+    static std::shared_ptr<ZarrArray> Create(const std::string& osParentName,
+                                             const std::string& osName,
+                                             const std::vector<std::shared_ptr<GDALDimension>>& aoDims,
+                                             const GDALExtendedDataType& oType,
+                                             const std::vector<DtypeElt>& aoDtypeElts,
+                                             const std::vector<GUInt64>& anBlockSize,
+                                             bool bFortranOrder);
+
+    bool IsWritable() const override { return false; }
+
+    const std::string& GetFilename() const override { return m_osFilename; }
+
+    const std::vector<std::shared_ptr<GDALDimension>>& GetDimensions() const override { return m_aoDims; }
+
+    const GDALExtendedDataType& GetDataType() const override { return m_oType; }
+
+    std::vector<GUInt64> GetBlockSize() const override { return m_anBlockSize; }
+
+    const void* GetRawNoDataValue() const override { return m_pabyNoData; }
+
+    void RegisterNoDataValue(const void*);
+
+    void SetFilename(const std::string& osFilename ) { m_osFilename = osFilename; }
+
+    void SetDimSeparator(const std::string& osDimSeparator) { m_osDimSeparator = osDimSeparator; }
+
+    void SetDecompressor(const CPLCompressor* psComp) { m_psDecompressor = psComp; }
+
+    void SetAttributes(const CPLJSONObject& attrs) { m_oAttrGroup.Init(attrs, m_bUpdatable); }
+
+    void SetSRS(const std::shared_ptr<OGRSpatialReference>& srs) { m_poSRS = srs; }
+
+    void SetRootDirectoryName(const std::string& osRootDirectoryName) { m_osRootDirectoryName = osRootDirectoryName; }
+
+    void SetVersion(int nVersion) { m_nVersion = nVersion; }
+
+    std::shared_ptr<GDALAttribute> GetAttribute(const std::string& osName) const override
+        { return m_oAttrGroup.GetAttribute(osName); }
+
+    std::vector<std::shared_ptr<GDALAttribute>> GetAttributes(CSLConstList papszOptions) const override
+        { return m_oAttrGroup.GetAttributes(papszOptions); }
+
+    std::shared_ptr<OGRSpatialReference> GetSpatialRef() const override;
+};
+
+#endif // ZARR_H
